Update repo docs and hierarchy to reflect demo intent
I removed the extra documentation around Consul NIA to emphasize
the simplicity of the module structure for compatible modules. I
have also changed the initial design to possibly have a central
repo that houses multiple modules or even "official" modules. These
will be later defined and clarified on its criterial. For now, we
are moving forward with simple modules that can be housed in the
author's namespace or organization.